# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DigiPulse API",
        "description": "API documentation for the DigiPulse monitoring system. **Authentication:** Use these endpoints to register, login, and manage user tokens.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/support/tickets": {
            "get": {
                "tags": [
                    "Support"
                ],
                "summary": "List user support tickets",
                "description": "List user's support tickets.",
                "operationId": "45c7216497473c3d2ca5989d12b2f67b",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SupportTicketSchema"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "frontendKey": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Support"
                ],
                "summary": "Create a new support ticket",
                "description": "Store a newly created support ticket.",
                "operationId": "382a8b3e11ed0a8f753d1470ccb467f6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "subject",
                                    "message"
                                ],
                                "properties": {
                                    "subject": {
                                        "type": "string",
                                        "example": "Issue with payment"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "I was charged twice."
                                    },
                                    "priority": {
                                        "type": "string",
                                        "example": "medium",
                                        "enum": [
                                            "low",
                                            "medium",
                                            "high"
                                        ]
                                    },
                                    "contact_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket created successfully"
                    }
                },
                "security": [
                    {
                        "frontendKey": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/support/tickets/{ticket}": {
            "get": {
                "tags": [
                    "Support"
                ],
                "summary": "Get specific support ticket",
                "description": "Display a specific support ticket with its messages.",
                "operationId": "2e94d10278e8777e73bf8b522a572841",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "description": "The ticket ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SupportTicketSchema"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "frontendKey": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/support/tickets/{ticket}/reply": {
            "post": {
                "tags": [
                    "Support"
                ],
                "summary": "Add a message to an existing ticket",
                "description": "Add a message to an existing ticket.",
                "operationId": "0f6785dcf727637816d81c0036a918bb",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "description": "The ticket ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Thank you for the help!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reply sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SupportTicketMessageSchema"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "frontendKey": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SupportTicketMessageSchema": {
                "title": "Support Ticket Message",
                "description": "A message within a support ticket",
                "required": [
                    "id",
                    "message",
                    "is_admin_reply"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "message": {
                        "type": "string",
                        "example": "Thank you for your request. We are looking into it."
                    },
                    "is_admin_reply": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user_name": {
                        "type": "string",
                        "example": "John Doe"
                    }
                },
                "type": "object"
            },
            "SupportTicketSchema": {
                "title": "Support Ticket",
                "description": "A support ticket record",
                "required": [
                    "id",
                    "subject",
                    "message",
                    "status",
                    "priority"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "subject": {
                        "type": "string",
                        "example": "Help with site setup"
                    },
                    "message": {
                        "type": "string",
                        "example": "I am having trouble adding my first site."
                    },
                    "status": {
                        "type": "string",
                        "example": "open",
                        "enum": [
                            "open",
                            "closed",
                            "pending"
                        ]
                    },
                    "priority": {
                        "type": "string",
                        "example": "medium",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ]
                    },
                    "contact_email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "messages": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SupportTicketMessageSchema"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "frontendKey": {
                "type": "apiKey",
                "description": "Custom frontend key for API access",
                "name": "X-Frontend-Key",
                "in": "header"
            },
            "bearerAuth": {
                "type": "http",
                "description": "BearerAuth access token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}